Chapter05/vision-api/requirements.txt
google-cloud-storage==1.7.0
google-cloud-vision==0.30.0
Pillow==5.0.0
